refactor: changed way internal logic works and the exposed api
