+++ packages/@tinacms/react-sidebar/src/sidebar.ts
@@ -17,13 +17,14 @@ limitations under the License.
 */
 
 import { EventBus, Callback } from '@tinacms/core'
+import { NoFormsPlaceholder } from './components/NoFormsPlaceHolder'
 import * as React from 'react'
 
 export interface SidebarStateOptions {
   hidden?: boolean
   position?: SidebarPosition
   buttons?: SidebarButtons
-  emptyState?: React.ElementType
+  placeholder?: React.ReactNode
 }
 
 export interface SidebarButtons {
@@ -35,7 +36,7 @@ export declare type SidebarPosition = 'fixed' | 'float' | 'displace' | 'overlay'
 
 export class SidebarState {
   private _isOpen: boolean = false
-  emptyState?: React.ElementType
+  placeholder: React.ReactNode
 
   position: SidebarPosition = 'displace'
   _hidden: boolean = false
@@ -47,7 +48,8 @@ export class SidebarState {
   constructor(private events: EventBus, options: SidebarStateOptions = {}) {
     this.position = options.position || 'displace'
     this._hidden = !!options.hidden
-    this.emptyState = options.emptyState
+    this.placeholder =
+      options.placeholder || React.createElement(NoFormsPlaceholder)
 
     if (options.buttons?.save) {
       this.buttons.save = options.buttons.save
