+++ packages/demo-next/pages/_app.js
@@ -31,7 +31,7 @@ export default class Site extends App {
     super()
     this.cms = new TinaCMS({
       sidebar: {
-        emptyState: Empty,
+        placeholder: <Empty />,
         position: 'overlay',
         hidden: process.env.NODE_ENV === 'production',
       },
